+++ staticsite/features/markdown.py
@@ -392,7 +392,7 @@ class MarkdownPage(Page):
             self.body_rest = None
 
     def check(self, checker):
-        self.mdpages.render_page(self)
+        self.content
 
     @lazy
     def content_short(self):